+++ src/extensions/rules/rulesext.h
@@ -196,6 +196,18 @@ public:
     VoxType DetectBeaconVoice;
 
     /**
+     *  Defines the game-wide cap (in percentages) that technos can self-heal.
+     *  This is the default used by technos that don't have this key explicitly specified for them.
+     */
+    float SelfHealingCap;
+
+    /**
+     *  Defines the game-wide rate (in minutes) that technos will self-heal.
+     *  This is the default used by technos that don't have this key explicitly specified for them.
+     */
+    float SelfHealingRate;
+
+    /**
      *  Is LandType Beach considered as "requires crushing" for passability purposes, as opposed to water?
      */
     bool IsBeachIsCrush;